backend/enrollment-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pyjwt>=2.8.0
httpx>=0.25.2